COMP: Avoid implicit narrowing casts
Explicitly cast initializers in {} array initializer where the assigned type is smaller than the value type. This avoids implicit narrowing in a {} initializer, which is ill-formed in C++11.

git-svn-id: http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment@17052 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Algorithm/vtkImageEMLocalSegmenter.cxx
@@ -692,7 +692,12 @@ static void vtkImageEMLocalSegmenterExecute(vtkImageEMLocalSegmenter *self,float
   // -----------------------------------------------------
   vtkIdType outIncX, outIncY, outIncZ;
   outData->GetContinuousIncrements(outExt, outIncX, outIncY, outIncZ);
-  int outInc[3] = {outIncX, outIncY, outIncZ};
+  int outInc[3] =
+    {
+    static_cast<int>(outIncX),
+    static_cast<int>(outIncY),
+    static_cast<int>(outIncZ)
+    };
   vtkImageEMLocalSegmenter_TransfereDataToOutputExtension(self,OutputVector,outPtr,outInc,0);
 
   delete[] OutputVector;
